SpectrumAnalyst in lib-fft
